+++ src/routes/Donations/index.ts
@@ -1,0 +1,7 @@
+import { Router } from 'express';
+
+import GetTopDonation from './GetTopDonation';
+
+const donations = Router();
+
+donations.get('/top', (req, res, next) => {});