+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## [0.0.3] - 2020-08-13
+
+### Added
+
+- Infura subscription for contract events
+- API routes
+- New contract ABI with log events included
+
 ## [0.0.2] - 2020-08-12
 
 ### Added